Adds cors and ingress
