+++ deploy/circle-api.ingress.yaml
@@ -1,3 +1,5 @@
+# kubetpl:syntax:$
+
 apiVersion: networking.k8s.io/v1
 kind: Ingress
 metadata:
@@ -8,8 +10,8 @@ metadata:
       {"Type": "redirect", "RedirectConfig": { "Protocol": "HTTPS", "Port":
       "443", "StatusCode": "HTTP_301"}}
     alb.ingress.kubernetes.io/certificate-arn: >-
-      arn:aws:acm:us-east-2:581679387567:certificate/a28a7d85-7d96-4f41-b83b-938135616b91
-    alb.ingress.kubernetes.io/group.name: wormscan-group-staging
+      $CERTIFICATE
+    alb.ingress.kubernetes.io/group.name: cctp-relayer
     alb.ingress.kubernetes.io/healthcheck-interval-seconds: '15'
     alb.ingress.kubernetes.io/healthcheck-path: /v1/health
     alb.ingress.kubernetes.io/healthcheck-port: traffic-port
@@ -22,11 +24,11 @@ metadata:
     alb.ingress.kubernetes.io/tags: createdBy=aws-controller
     alb.ingress.kubernetes.io/target-type: ip
     alb.ingress.kubernetes.io/unhealthy-threshold-count: '2'
-    external-dns.alpha.kubernetes.io/hostname: $HOST
+    external-dns.alpha.kubernetes.io/hostname: $HOSTNAME
     kubernetes.io/ingress.class: alb
 spec:
   rules:
-    - host: $HOST
+    - host: $HOSTNAME
       http:
         paths:
           - path: /
@@ -35,4 +37,4 @@ spec:
               service:
                 name: circle-api
                 port:
-                  number: 80
+                  number: 8080
