+++ src/api/index.ts
@@ -1,9 +1,11 @@
 import Koa from "koa";
 import Router from "koa-router";
+import cors from "@koa/cors";
 import { Relay } from "../data/relay.model";
 import { setupDb } from "../data/db";
 import { getLogger } from "../common/logging";
 import { RelayController } from "./relays.controller";
+import { healthMiddleware, loggerMiddleware } from "./middleware";
 
 const config = {
   env: process.env.NODE_ENV || "local",
@@ -19,31 +21,21 @@ const config = {
 async function main() {
   const logger = getLogger(config.env, config.logLevel);
   const app = new Koa();
+  logger.info("connecting to db...");
   await setupDb(config.db);
+  logger.info("connected to db.");
   const relayCtrl = new RelayController(Relay);
 
   const router = new Router();
   router.prefix("/v1");
+  // Admin endpoints
+  router.get("/health", healthMiddleware);
 
+  // Relay endpoints
   router.get("/relays", relayCtrl.search);
 
-  router.get("/health", async (ctx) => {
-    ctx.body = "ok";
-  });
-
-  app.use(async (ctx, next) => {
-    logger.debug(`${ctx.method} ${ctx.url}`);
-    const start = process.hrtime();
-    await next();
-    const [_, inNanos] = process.hrtime(start);
-    const duration = inNanos / 1e6;
-    logger.debug(`${ctx.method} ${ctx.url} ${ctx.status} ${duration}ms`, {
-      method: ctx.method,
-      url: ctx.url,
-      duration,
-      status: ctx.status,
-    });
-  });
+  app.use(cors({ origin: "*" }));
+  app.use(loggerMiddleware(logger));
   app.use(router.allowedMethods());
   app.use(router.middleware());
