+++ src/package.json
@@ -1,6 +1,6 @@
 {
   "name": "src",
-  "version": "0.1.3",
+  "version": "0.1.6",
   "description": "",
   "main": "index.js",
   "scripts": {
@@ -26,6 +26,7 @@
     "@certusone/wormhole-sdk": "^0.9.14",
     "@improbable-eng/grpc-web-node-http-transport": "^0.15.0",
     "@influxdata/influxdb-client": "^1.33.2",
+    "@koa/cors": "^4.0.0",
     "@wormhole-foundation/relayer-engine": "github:wormhole-foundation/relayer-engine#7698ed65ccce01e3bf692b978bfb4a9ffd3e5514",
     "ethers": "^5.7.2",
     "ioredis": "^5.3.1",
@@ -40,6 +41,7 @@
   "license": "ISC",
   "devDependencies": {
     "@types/koa": "^2.13.6",
+    "@types/koa__cors": "^4.0.0",
     "@types/koa-router": "^7.4.4",
     "prettier": "^2.8.6"
   }
